Better bashrc and inputrc
Some defaults added from raspberry pi. I want to keep the pi as simple
as possible, so I'll just stick to the default shell for a while and see
how good I can get that. If I would be going for zsh, I would
immediately want oh-my-zsh and things get a "bloatier" feeling quickly.

Adding an inputrc too since that changes defaults for readline, which
should have an effect in more places than just bash.

Finally actually adding a bash_aliases file since it is the cleaner way
to go at it.